backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
chromadb>=0.5.5
sentence-transformers>=3.0.1
transformers>=4.44

pydantic>=2.7
google-generativeai>=0.7.0
python-dotenv>=1.0
tqdm>=4.66

pydantic-settings>=2.2
